Conflicts resolved:
source/blender/blenloader/intern/readfile.c

+++ source/blender/blenlib/intern/storage.c
@@ -469,7 +469,8 @@ int BLI_exist(const char *name)
 }
 
 /* would be better in fileops.c except that it needs stat.h so add here */
-int BLI_is_dir(const char *file) {
+int BLI_is_dir(const char *file)
+{
 	return S_ISDIR(BLI_exist(file));
 }
 
@@ -478,17 +479,17 @@ LinkNode *BLI_read_file_as_lines(const char *name)
 	FILE *fp= fopen(name, "r");
 	LinkNode *lines= NULL;
 	char *buf;
-	int64_t size;
+	size_t size;
 
 	if (!fp) return NULL;
 		
 	fseek(fp, 0, SEEK_END);
-	size= ftell(fp);
+	size= (size_t)ftell(fp);
 	fseek(fp, 0, SEEK_SET);
 
 	buf= MEM_mallocN(size, "file_as_lines");
 	if (buf) {
-		int i, last= 0;
+		size_t i, last= 0;
 		
 			/* 
 			 * size = because on win32 reading
